Added expire and stop plugins to redis monitor
This should complete the functionality for now for the redis monitor plugins. Moved both the 'stop' functionality and the 'expire' functionality into two separate plugins that you can turn on/off as needed.

Also modded the kafkadump utility to support either reading from the beginning of the topic or from the end
